fix: Failed when primary string had repeated letters
Store the count of occurrence of the letters as value in dictionary

The conversion of string to a dictionary of the individual characters ignore the repeated characters in the primary string. So the test case failed for 'abhiraj' and 'aa', despite 'abhiraj' having 'aa'. Change was done so that the count was stored as the value to the key.